requirements.txt
# --------- torch --------- #
torch>=2.0.0
transformers>=4.32.0
lightning>=2.0.0

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0

# --------- logger -------- #
wandb>=0.10.0

# --------- others -------- #
rootutils~=1.0.7
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
omegaconf>=2.3.0
datasets==2.14.5
pydantic>=1.8.2
tqdm>=4.62.3
